Made panes load before they are meant to appear on the screen.

+++ deco2800.arcade.hunter/src/main/java/deco2800/arcade/hunter/screens/GameScreen.java
@@ -66,8 +66,7 @@ public class GameScreen implements Screen {
 		backgroundLayer = new BackgroundLayer(0);
 		spriteLayer = new SpriteLayer((float) 0.6);
 
-		int numPanes = (int) (Math.ceil(Config.screenWidth
-				/ Config.PANE_SIZE_PX) + 1);
+		int numPanes = (int) (Math.ceil(Config.screenWidth / (double) Config.PANE_SIZE_PX) + 1);
 		foregroundLayer = new ForegroundLayer(1, numPanes);
 
 		// Spawn entities
